+++ CHANGELOG.md
@@ -5,6 +5,10 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Added
+
+- database: add GetAllManagedDatabases to retrieve all managed databases regardless of paging
+
 ## [8.27.0]
 
 ### Added